Move private modules into a `-private` directory.
This makes it clear that these are not intended to be public API...
