+++ addon/-private/engine-ext.js
