+++ addon/-private/engine-instance-ext.js
