+++ addon/-private/ext-require.js
