+++ addon/-private/link-to-component-ext.js
