+++ addon/-private/router-dsl-ext.js
