+++ addon/-private/router-ext.js
